requirements.txt
# Core dependencies for Holographic Memory System with Emotion Engine
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pyyaml>=6.0

# Optional dependencies for enhanced functionality
matplotlib>=3.5.0  # For progress visualization
requests>=2.25.0   # For HTTP requests (Ollama integration)

# Development dependencies (optional)
pytest>=6.0.0      # For running tests
black>=21.0.0      # Code formatting
isort>=5.0.0       # Import sorting
flake8>=4.0.0      # Linting

# Documentation dependencies (optional)
sphinx>=4.0.0      # Documentation generation
sphinx-rtd-theme>=1.0.0  # Documentation theme